Update component and folder structure
Also updates the EuiCacheProvider component
