+++ src/app/components/icons/connector_icons.ts
