+++ src/app/components/index.tsx
@@ -1,8 +1,12 @@
 import { SideNav } from "./SideNav";
 import { SplitPanel } from "./SplitPanel";
 import { IndicesTable } from "./IndicesTable";
+import { AppProvider } from "./AppProvider";
+import { PageWrapper } from "./PageWrapper";
 //
 export {
+  AppProvider,
+  PageWrapper,
   SideNav,
   SplitPanel,
   IndicesTable,
