+++ src/app/layout.tsx
@@ -1,18 +1,18 @@
-'use client';
 // Main Wrapping Layout
 // Includes Top Bars and Sidenav
 //
-import React from 'react';
+import React, { Suspense } from 'react';
 //
+import '../lib/polyfills';
 import '@elastic/eui/dist/eui_theme_light.css';
 import {
   EuiHeader,
   EuiHeaderLogo,
   EuiPageTemplate
 } from '@elastic/eui';
-import { EmotionCacheProvider } from '@/utils/emotion-cache-provider';
-import { CustomEuiProvider } from '@/utils/custom-eui-provider';
-import { SideNav } from './components';
+
+import { AppProvider } from '@/components/AppProvider';
+import { PageWrapper } from '@/components/PageWrapper';
 //
 //
 export default function Layout({
@@ -23,32 +23,11 @@ export default function Layout({
   return (
     <html lang="en">
       <body>
-        <EmotionCacheProvider
-          options={{ key: 'search-next' }}
-          providerCacheKey="cache"
-          providerComponent={CustomEuiProvider}
-        >
-          <EuiHeader
-            theme="dark"
-            sections={[
-              {
-                items: [
-                  <EuiHeaderLogo key="elastic">Elastic</EuiHeaderLogo>
-                ]
-              }
-            ]}
-          />
-          <EuiHeader />
-          <EuiPageTemplate
-            restrictWidth={false}
-            panelled={true}
-          >
-            <EuiPageTemplate.Sidebar sticky>
-              <SideNav />
-            </EuiPageTemplate.Sidebar>
-            {children}
-          </EuiPageTemplate>
-        </EmotionCacheProvider>
+        <AppProvider>
+          <Suspense>
+            <PageWrapper>{children}</PageWrapper>
+          </Suspense>
+        </AppProvider>
       </body>
     </html>
   );
